Fixup flag parsing precedence / string list gathering.
# Building wheels and fs_util will be skipped. Delete if not intended.
[ci skip-build-wheels]
