+++ src/rust/engine/client/src/options/args.rs
@@ -6,6 +6,8 @@ use std::env;
 use super::id::{NameTransform, OptionId, Scope};
 use super::parse::parse_bool;
 use super::OptionsSource;
+use crate::options::parse::parse_string_list;
+use crate::options::ListEdit;
 
 pub(crate) struct Args {
   args: Vec<String>,
@@ -39,7 +41,7 @@ impl Args {
   }
 
   fn find_string(&self, arg_name: &str) -> Result<Option<String>, String> {
-    for arg in &self.args {
+    for arg in self.args.iter().rev() {
       let mut components = arg.as_str().splitn(2, '=');
       if let Some(name) = components.next() {
         if name == arg_name {
@@ -59,7 +61,7 @@ impl OptionsSource for Args {
   fn get_string(&self, id: &OptionId) -> Result<Option<String>, String> {
     if let Some(switch) = id.2 {
       let prefixes = [format!("-{}=", switch), format!("-{}", switch)];
-      for arg in &self.args {
+      for arg in self.args.iter().rev() {
         for prefix in &prefixes {
           if arg.starts_with(&*prefix) {
             return Ok(Some(arg[prefix.len()..].to_owned()));
@@ -85,6 +87,30 @@ impl OptionsSource for Args {
           None => Ok(None),
         }
       }
+    }
+  }
+
+  fn get_string_list(&self, id: &OptionId) -> Result<Option<Vec<ListEdit<String>>>, String> {
+    let arg_name = Self::arg_name(id, Negate::FALSE);
+    let mut edits = vec![];
+    for arg in &self.args {
+      let mut components = arg.as_str().splitn(2, '=');
+      if let Some(name) = components.next() {
+        if name == arg_name {
+          let value = components.next().ok_or_else(|| {
+            format!(
+              "Expected string list option {name} to have a value.",
+              name = name
+            )
+          })?;
+          edits.extend(parse_string_list(name, value)?)
+        }
+      }
+    }
+    if edits.is_empty() {
+      Ok(None)
+    } else {
+      Ok(Some(edits))
     }
   }
 }
